型定義を本体ファイルに追加→d.tsファイルを削除
